+++ src/scripts/marcRecordUi.js
@@ -3,7 +3,7 @@
 /* eslint-disable max-statements */
 
 import {activateEditorButtons} from './editorButtons.js';
-import {isDataFieldTag} from './editorUtils.js';
+import {isDataFieldTa, stringToMarcField} from './editorUtils.js';
 
 //****************************************************************************//
 //                                                                            //
@@ -108,93 +108,6 @@ export function getEditorFields(editorElementId = 'Record', subfieldCodePrefix =
     return [];
   }
   return [...parentElem.children].map(div => stringToMarcField(div.textContent, subfieldCodePrefix)); // [].map() converts children into an editable array
-}
-
-export function stringToMarcField(str, subfieldCodePrefix = '$$') { // export since used in tests. settings.subfieldCodePrefix
-  // console.log(`String2field: '${str}', '${subfieldCodePrefix}'`);
-  const len = str.length;
-  if (len <= 3) {
-    return {tag: str, error: `Incomplete field ${str}`};
-  }
-
-  const tag = str.substring(0, 3);
-
-  const ind1 = normalizeIndicator(str.substring(3, 4), tag);
-  if ( len === 4) {
-    return {tag, ind1, error: `Incomplete field ${tag}`}; //, ind2: ' '};
-  }
-
-  const ind2 = normalizeIndicator(str.substring(4, 5), tag);
-  if ( len === 5) {
-    return {tag, ind1, ind2, error: `Incomplete field ${tag}`};
-  }
-
-  const rest = str.substring(5);
-  if (!isDataFieldTag(tag)) {
-    return {tag, ind1, ind2, value: rest.replace(/#/gu, ' '), error: rest.length <= 5 ? `Incomplete field ${tag}` : false};
-  }
-
-  const {subfields, error} = convertDataToSubfields(tag, rest, subfieldCodePrefix);
-  //if (subfields.length === 0) {
-  if (error) {
-    //console.log(`Failed to parse '${str}': ${error}`);
-    return {tag, ind1, ind2, value: rest, error: `${tag}: ${error}`}; // This is erronous state, as subfields failed to parse
-  }
-
-  return {tag, ind1, ind2, subfields, error: false};
-
-  function normalizeIndicator(ind, tag) { // convert data from web page to marc
-    //console.log(`Process indicator '${ind}'`);
-    if (!isDataFieldTag(tag)) {
-        // Even control fields contain slots/fake indicators for indicators, so that records looks good in the browser.
-        // Fake indicators are removed when fields are converted into a real record.
-        return ' ';
-    }
-    if ( ind.match(/^[0-9]$/u) ) {
-        return ind;
-    }
-    // Note: this converts illegal indicator values to default value '#'. Needed by editor
-    return ' ';
-  }
-
-}
-
-function convertDataToSubfields(tag, data, separator = '$$') {
-  if (separator.length < 1) {
-    return {subfields: [], error: 'Missing subfield separator string'};
-  }
-  if ( data.length < 4 ) {
-    return {subfields: [], error: 'Not enough data yet'};
-  }
-  if ( data.substring(0, separator.length) !== separator ) {
-    return {subfields: [], error: `Data segment should begin with '${separator}'`};
-  }
-  const data2 = data.substring(separator.length);
-  const subfields = data2.split(separator);
-
-  const noSubfieldCodeIndex = subfields.findIndex((sf, i) => sf.length < 1); // 1st char is code and the rest is data
-  if (noSubfieldCodeIndex > -1) {
-    return{subfields: [], error: `Subfield #${noSubfieldCodeIndex+1} does not contain a subfield code (nor data)`};
-  }
-
-
-  const illegalSubfieldCodeIndex = subfields.findIndex((sf, i) => sf.match(/^[^a-z0-9]/u));
-  if (illegalSubfieldCodeIndex > -1) {
-    return {subfields: [], error: `Subfield #${illegalSubfieldCodeIndex+1} has unexpected subfield code '${subfields[illegalSubfieldCodeIndex].substring(0, 1)}'`};
-  }
-
-  if (tag !== 'CAT') { // CAT's empty $b is so common, that there's no point to complain about it, esp. as it is oft uneditable.
-    const emptySubfieldIndex = subfields.findIndex((sf, i) => sf.length < 2); // 1st char is code and the rest is data
-    if (emptySubfieldIndex > -1) {
-      return{subfields: [], error: `Subfield #${emptySubfieldIndex+1} (${separator}${subfields[emptySubfieldIndex].substring(0, 1)}) does not contain any data`};
-    }
-  }
-
-  return { subfields: subfields.map(sf => stringToSubfield(sf)), error: undefined};
-
-  function stringToSubfield(str) {
-    return {'code': str.substring(0, 1), 'value': str.substring(1)};
-  }
 }
